Override of isInRangeToRenderDist on flying aggressive entity. This allows flying creatures to render even when far away or high up.
First commit of my first attempt at animating the Therizinosaurus. It is NOT done at all, but at least the legs move and head and tail bob.

Also, you may see me commit just so I can work on multiple computers and need to keep them in sync. Will probably do this twice a day. Let me know if that's a problem.